fix: renamedSliderImageCard back to MainCard
